Added ability to change the speed of a transponder, added the ability to give integer values in commnads, fixed bug adding the ability to specify doubles instead of only integers for values

+++ src/main/java/OTN/Commands/Tokens/Token.java
@@ -26,7 +26,7 @@ public class Token{
 
     public types type;
     public String value;
-    public int intValue;
+    public double intValue;
 
     public Token(types type, String value){
 
@@ -34,10 +34,11 @@ public class Token{
         this.value = value;
     }
 
-    public Token(int value){
+    public Token(double value){
 
         type = types.INT;
         intValue = value;
 
     }
+
 }
